requirements.txt
requests>=2.28.0
PyYAML>=6.0
pathlib2>=2.3.7; python_version < "3.4"